[Feat] BaseEntity 추가 - model 폴더 변경 - BaseEntity 상속
