+++ backend/src/main/java/com/woof/api/orders/model/entity/OrdersMyList.java
@@ -1,4 +1,4 @@
-package com.woof.api.orders.model.dto;
+package com.woof.api.orders.model.entity;
 
 import lombok.*;
