+++ backend/src/main/java/com/woof/api/orders/model/response/OrdersListRes2.java
@@ -1,4 +1,4 @@
-package com.woof.api.orders.model.dto;
+package com.woof.api.orders.model.response;
 
 import lombok.Builder;
 import lombok.Data;
@@ -7,11 +7,10 @@ import lombok.Getter;
 @Builder
 @Data
 @Getter
-public class OrdersReadRes {
+public class OrdersListRes2 {
     private Long idx;
     private String name;
     private String phoneNumber;
     private Integer time;
     private String place;
-    private String reservation_status;
 }
