+++ backend/src/main/java/com/woof/api/orders/model/response/OrdersReadRes.java
@@ -1,4 +1,4 @@
-package com.woof.api.orders.model.dto;
+package com.woof.api.orders.model.response;
 
 import lombok.Builder;
 import lombok.Data;
@@ -7,10 +7,11 @@ import lombok.Getter;
 @Builder
 @Data
 @Getter
-public class OrdersListRes2 {
+public class OrdersReadRes {
     private Long idx;
     private String name;
     private String phoneNumber;
     private Integer time;
     private String place;
+    private String reservation_status;
 }
