+++ backend/src/main/java/com/woof/api/orders/model/response/OrdersReadRes2.java
@@ -1,16 +1,17 @@
-package com.woof.api.orders.model.dto;
+package com.woof.api.orders.model.response;
 
 import lombok.Builder;
 import lombok.Data;
 
-import java.util.List;
 
 @Builder
 @Data
-public class OrdersListRes {
+public class OrdersReadRes2 {
     private Boolean isSuccess;
     private Integer code;
     private String message;
-    private List<OrdersReadRes> result;
+    private OrdersReadRes result;
     private Boolean success;
+
+
 }
