+++ backend/src/main/java/com/woof/api/orders/repository/OrderRepository.java
@@ -2,11 +2,9 @@ package com.woof.api.orders.repository;
 
 
 
-import com.woof.api.orders.model.Orders;
-import com.woof.api.orders.model.dto.OrderDto;
-import com.woof.api.orders.model.dto.OrdersMyList;
+import com.woof.api.orders.model.entity.Orders;
+import com.woof.api.orders.model.entity.OrderDto;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 import java.util.Optional;
