+++ backend/src/main/java/com/woof/api/payment/service/PaymentService.java
@@ -1,32 +1,22 @@
 package com.woof.api.payment.service;
 
-import com.google.gson.Gson;
-import com.google.gson.JsonObject;
 import com.siot.IamportRestClient.IamportClient;
 import com.siot.IamportRestClient.exception.IamportResponseException;
 import com.siot.IamportRestClient.request.AgainPaymentData;
-import com.siot.IamportRestClient.request.CancelData;
-import com.siot.IamportRestClient.request.ScheduleEntry;
 import com.siot.IamportRestClient.request.UnscheduleData;
 import com.siot.IamportRestClient.response.IamportResponse;
 import com.siot.IamportRestClient.response.Payment;
 import com.siot.IamportRestClient.response.Schedule;
-import com.woof.api.orders.model.Orders;
 import com.woof.api.payment.model.SubscribeInfo;
-import com.woof.api.payment.model.dto.response.GetPortOneRes;
 import lombok.RequiredArgsConstructor;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
-import javax.net.ssl.HttpsURLConnection;
 import java.io.*;
 import java.math.BigDecimal;
-import java.net.URL;
-import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
-import java.util.Map;
 
 
 @Service
