+++ backend/src/main/java/com/woof/api/review/model/entity/Review.java
@@ -2,8 +2,6 @@ package com.woof.api.review.model.entity;
 
 
 import com.woof.api.common.BaseEntity;
-import com.woof.api.orders.model.Orders;
-import com.woof.api.product.model.entity.ProductImage;
 import com.woof.api.product.model.entity.ProductSchool;
 import com.woof.api.product.model.entity.ProductManager;
 import lombok.*;
